Move tests into a fragment package
