+++ app/src/androidTest/java/com/github/sdpteam15/polyevents/fragments/MoreFragmentTest.kt
@@ -1,4 +1,4 @@
-package com.github.sdpteam15.polyevents
+package com.github.sdpteam15.polyevents.fragments
 
 
 import androidx.test.espresso.Espresso.onView
@@ -9,6 +9,10 @@ import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.sdpteam15.polyevents.ItemRequestActivity
+import com.github.sdpteam15.polyevents.ItemsAdminActivity
+import com.github.sdpteam15.polyevents.MainActivity
+import com.github.sdpteam15.polyevents.R
 import com.github.sdpteam15.polyevents.database.Database
 import com.github.sdpteam15.polyevents.fakedatabase.FakeDatabase
 import com.github.sdpteam15.polyevents.fakedatabase.FakeDatabaseItem
@@ -48,7 +52,7 @@ class MoreFragmentTest {
 
     @Test
     fun itemsAdminActivity() {
-        var availableItems: MutableMap<Item, Int> = mutableMapOf()
+        val availableItems: MutableMap<Item, Int> = mutableMapOf()
         availableItems[Item(null, "Chocolat", ItemType.OTHER)] = 30
         availableItems[Item(null, "Kiwis", ItemType.OTHER)] = 10
         availableItems[Item(null, "230V Plugs", ItemType.PLUG)] = 30
